fix price combo product trans

+++ sale-service/src/main/java/vn/viettel/sale/entities/ComboProductTransDetail.java
@@ -35,10 +35,10 @@ public class ComboProductTransDetail extends BaseEntity {
     @Column(name = "QUANTITY")
     private Integer quantity;
     @Column(name = "PRICE")
-    private Float price;
+    private Double price;
     @Column(name = "PRICE_NOT_VAT")
-    private Float priceNotVat;
+    private Double priceNotVat;
     @Column(name = "AMOUNT")
-    private Float amount;
+    private Double amount;
 
 }
